[Refactor] Aladin 도메인 Book 도메인에 병합 및 Book 도메인 Converter 분리
